requirements.txt
discord.py
python.env
python-dotenv
asyncio
os
